# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get full booking detail",
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/bookings/{id}/cancel": {
      "post": {
        "operationId": "BookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a confirmed booking",
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/admin/bookings": {
      "get": {
        "operationId": "AdminBookingsController_list",
        "parameters": [
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Filters on checkInAt (default: 7 days ago)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Filters on checkInAt (default: now)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "propertyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "hourly",
                "fullday"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Comma-separated list of BookingStatus values",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amountMin",
            "required": false,
            "in": "query",
            "description": "Minimum totalAmountPaise",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "amountMax",
            "required": false,
            "in": "query",
            "description": "Maximum totalAmountPaise",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "guestPhone",
            "required": false,
            "in": "query",
            "description": "Partial match on guest phone, min 3 chars",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingRef",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "enum": [
                "createdAt",
                "checkInAt",
                "checkOutAt",
                "totalAmountPaise"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List bookings with admin filters",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/kpis": {
      "get": {
        "operationId": "AdminBookingsController_kpis",
        "parameters": [
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "propertyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "hourly",
                "fullday"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amountMin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "amountMax",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "guestPhone",
            "required": false,
            "in": "query",
            "description": "Partial match on guest phone, min 3 chars",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingRef",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "KPI summary for the active filter set",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/active": {
      "get": {
        "operationId": "AdminBookingsController_active",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Currently checked-in bookings, soonest-expiring first",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/{id}": {
      "get": {
        "operationId": "AdminBookingsController_getDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Full booking detail incl. timeline and admin actions",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/{id}/void": {
      "post": {
        "operationId": "AdminBookingsController_void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Void a booking (fraud/error, auto-refunds 100%)",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/{id}/cancel": {
      "post": {
        "operationId": "AdminBookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCancelBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin-initiated cancellation",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/{id}/refund": {
      "post": {
        "operationId": "AdminBookingsController_refund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Issue a full or partial refund",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/{id}/force-checkout": {
      "post": {
        "operationId": "AdminBookingsController_forceCheckout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForceCheckoutBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Force-checkout a checked-in guest",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/{id}/extend": {
      "post": {
        "operationId": "AdminBookingsController_extend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Extend a checked-in booking checkout time",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/{id}/flag": {
      "post": {
        "operationId": "AdminBookingsController_flag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Flag a booking for review",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
    "/admin/bookings/{id}/unflag": {
      "post": {
        "operationId": "AdminBookingsController_unflag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a flag from a booking",
        "tags": [
          "Admin / Bookings"
        ]
      }
    }
  },
  "info": {
    "title": "QuickNest API",
    "description": "Backend APIs for QuickNest hotel booking platform",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CancelBookingDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Guest-supplied cancellation reason"
          }
        }
      },
      "VoidBookingDto": {
        "type": "object",
        "properties": {
          "reasonCategory": {
            "type": "string",
            "enum": [
              "fraud",
              "emergency",
              "duplicate",
              "hotel_request",
              "other"
            ]
          },
          "reasonText": {
            "type": "string"
          }
        },
        "required": [
          "reasonCategory"
        ]
      },
      "AdminCancelBookingDto": {
        "type": "object",
        "properties": {
          "reasonCategory": {
            "type": "string",
            "enum": [
              "guest_request",
              "hotel_request",
              "admin_decision",
              "system_error"
            ]
          },
          "reasonText": {
            "type": "string"
          }
        },
        "required": [
          "reasonCategory"
        ]
      },
      "RefundBookingDto": {
        "type": "object",
        "properties": {
          "amountPaise": {
            "type": "number",
            "description": "Amount to refund, in paise"
          },
          "reasonCategory": {
            "type": "string"
          },
          "reasonText": {
            "type": "string"
          }
        },
        "required": [
          "amountPaise",
          "reasonCategory"
        ]
      },
      "ForceCheckoutBookingDto": {
        "type": "object",
        "properties": {
          "reasonText": {
            "type": "string"
          }
        },
        "required": [
          "reasonText"
        ]
      },
      "ExtendBookingDto": {
        "type": "object",
        "properties": {
          "newCheckOutAt": {
            "type": "string",
            "description": "New check-out timestamp (ISO 8601)"
          },
          "extensionAmountPaise": {
            "type": "number",
            "description": "Extension surcharge, in paise"
          }
        },
        "required": [
          "newCheckOutAt",
          "extensionAmountPaise"
        ]
      },
      "FlagBookingDto": {
        "type": "object",
        "properties": {
          "flagType": {
            "type": "string",
            "enum": [
              "suspicious",
              "quality_issue",
              "partner_complaint",
              "other"
            ]
          },
          "flagNotes": {
            "type": "string"
          }
        },
        "required": [
          "flagType"
        ]
      }
    }
  }
}